+++ extension-al-object-designer/src/Commands/BrowserPreviewCommand.ts
@@ -10,12 +10,12 @@ export class BrowserPreviewCommand extends ALCommandBase {
     }
 
     async execute(message: any) {
-        let checkExt = vscode.extensions.getExtension('browser-preview');
+        let checkExt = vscode.extensions.getExtension('auchenberg.vscode-browser-preview');
         if (!checkExt) {
             await vscode.window.showErrorMessage(`Browser Preview extension is not installed or disabled.`);
         }
 
-        if (['Table','Page','PageExtension','TableExtension','Report', 'XmlPort', 'Query'].indexOf(message.Type) == -1) {
+        if (['Table','Page','PageExtension','TableExtension','Report', 'Query'].indexOf(message.Type) == -1) {
             await vscode.window.showErrorMessage(`${message.Type} objects cannot be run in Modern Client.`)
             return;
         }